flipit-simulation-master/requirements_drl.txt
# 深度强化学习相关依赖
torch>=1.9.0
torchvision>=0.10.0
torch-audio>=0.9.0
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0
pandas>=1.3.0
gymnasium>=0.29.1
gym>=0.21.0  # legacy-only pipeline
pyyaml>=5.4.0
tensorboard>=2.6.0
tqdm>=4.62.0

# 高级RL算法库（可选）
# stable-baselines3>=1.6.0
# ray[rllib]>=1.13.0

# 多智能体强化学习
pettingzoo>=1.12.0
supersuit>=3.4.0

# 神经网络可视化和分析
wandb>=0.12.0
plotly>=5.0.0

# 数据处理和优化
scikit-learn>=1.0.0
optuna>=2.10.0

